+++ tests/crossSections/blockF_WW_dilep.lua
@@ -1,19 +1,5 @@
-function append(t1, t2)
-    for i = 1, #t2 do
-        t1[#t1 + 1] = t2[i]
-    end
-
-    return t1
-end
-
-function copy_and_append(t1, t2)
-    local t3 = {}
-
-    append(t3, t1)
-    append(t3, t2)
-
-    return t3
-end
+local p1 = declare_input("p1")
+local p2 = declare_input("p2")
 
 parameters = {
     energy = 13000.,
@@ -25,21 +11,21 @@ cuba = {
     verbosity = 3,
     max_eval = 200000000,
     relative_accuracy = 0.005,
-    n_start = 10000000,   
-    seed = 5468730,        
+    n_start = 10000000,
+    seed = 5468730,
 }
 
 -- 'Flat' transfer functions to integrate over the visible particle's energies and angles
 -- First |P|
 FlatTransferFunctionOnP.tf_p_1 = {
     ps_point = add_dimension(),
-    reco_particle = 'input::particles/1',
+    reco_particle = p1.reco_p4,
     min = 0.,
     max = parameters.energy/2,
 }
 FlatTransferFunctionOnP.tf_p_2 = {
     ps_point = add_dimension(),
-    reco_particle = 'input::particles/2',
+    reco_particle = p2.reco_p4,
     min = 0.,
     max = parameters.energy/2,
 }
@@ -54,7 +40,7 @@ FlatTransferFunctionOnPhi.tf_phi_2 = {
     reco_particle = 'tf_p_2::output',
 }
 
--- Finally, do Theta 
+-- Finally, do Theta
 FlatTransferFunctionOnTheta.tf_theta_1 = {
     ps_point = add_dimension(),
     reco_particle = 'tf_phi_1::output',
@@ -102,13 +88,13 @@ Looper.looper = {
 }
 
     full_inputs = copy_and_append(inputs, {'looper::particles/1', 'looper::particles/2'})
-    
+
     BuildInitialState.initial_state = {
         particles = full_inputs
     }
 
     jacobians = {
-      'tf_p_1::TF_times_jacobian', 'tf_p_2::TF_times_jacobian', 
+      'tf_p_1::TF_times_jacobian', 'tf_p_2::TF_times_jacobian',
       'tf_phi_1::TF_times_jacobian', 'tf_phi_2::TF_times_jacobian',
       'tf_theta_1::TF_times_jacobian', 'tf_theta_2::TF_times_jacobian',
       'flatter_s13::jacobian', 'flatter_s24::jacobian',
@@ -138,7 +124,7 @@ Looper.looper = {
             pdg_id = 13,
             me_index = 3,
           },
-          
+
           {
             pdg_id = 12,
             me_index = 2,