+++ tests/phaseSpaceVolume/blockA.lua
@@ -1,3 +1,7 @@
+local p1 = declare_input("p1")
+local p2 = declare_input("p2")
+local p3 = declare_input("p3")
+
 load_modules('MatrixElements/dummy/libme_dummy.so')
 
 parameters = {
@@ -8,15 +12,15 @@ cuba = {
     verbosity = 3,
     max_eval = 200000000,
     relative_accuracy = 0.001,
-    n_start = 20000000,   
-    seed = 5468960,        
+    n_start = 20000000,
+    seed = 5468960,
 }
 
 -- 'Flat' transfer functions to integrate over the visible particle's energies and angles
 -- First |P|: only the input not concerned by blockA is integrated over
 FlatTransferFunctionOnP.tf_p_3 = {
     ps_point = add_dimension(),
-    reco_particle = 'input::particles/3',
+    reco_particle = p3.reco_p4,
     min = 0.,
     max = parameters.energy/2,
 }
@@ -24,18 +28,18 @@ FlatTransferFunctionOnP.tf_p_3 = {
 -- Pass these outputs over for Phi: the two outputs of blockA are integrated over their angles
 FlatTransferFunctionOnPhi.tf_phi_1 = {
     ps_point = add_dimension(),
-    reco_particle = 'input::particles/1',
+    reco_particle = p1.reco_p4,
 }
 FlatTransferFunctionOnPhi.tf_phi_2 = {
     ps_point = add_dimension(),
-    reco_particle = 'input::particles/2',
+    reco_particle = p2.reco_p4,
 }
 FlatTransferFunctionOnPhi.tf_phi_3 = {
     ps_point = add_dimension(),
     reco_particle = 'tf_p_3::output',
 }
 
--- Finally, do Theta 
+-- Finally, do Theta
 FlatTransferFunctionOnTheta.tf_theta_1 = {
     ps_point = add_dimension(),
     reco_particle = 'tf_phi_1::output',
@@ -79,8 +83,8 @@ Looper.looper = {
 
     jacobians = {
       'tf_p_3::TF_times_jacobian',
-      'tf_phi_1::TF_times_jacobian', 'tf_phi_2::TF_times_jacobian', 'tf_phi_3::TF_times_jacobian', 
-      'tf_theta_1::TF_times_jacobian', 'tf_theta_2::TF_times_jacobian', 'tf_theta_3::TF_times_jacobian', 
+      'tf_phi_1::TF_times_jacobian', 'tf_phi_2::TF_times_jacobian', 'tf_phi_3::TF_times_jacobian',
+      'tf_theta_1::TF_times_jacobian', 'tf_theta_2::TF_times_jacobian', 'tf_theta_3::TF_times_jacobian',
       'looper::jacobian', 'phaseSpaceOut::phase_space'
     }
 
@@ -89,7 +93,7 @@ Looper.looper = {
 
       matrix_element = 'dummy_matrix_element',
       matrix_element_parameters = {},
-      
+
       initialState = 'initial_state::partons',
 
       particles = {